+++ lib/glimmer/web/element_proxy.rb
@@ -465,6 +465,19 @@ module Glimmer
         end
       end
       
+      # Data-binds the generation of nested content to a model/property (in binding args)
+      # consider providing an option to avoid initial rendering without any changes happening
+      def bind_content(*binding_args, &content_block)
+        # TODO in the future, consider optimizing code by diffing content if that makes sense
+        content_binding_work = proc do |*values|
+          children.dup.each { |child| child.remove }
+          content(&content_block)
+        end
+        content_binding_observer = Glimmer::DataBinding::Observer.proc(&content_binding_work)
+        content_binding_observer.observe(*binding_args)
+        content_binding_work.call # TODO inspect if we need to pass args here (from observed attributes) [but it's simpler not to pass anything at first]
+      end
+      
       def respond_to_missing?(method_name, include_private = false)
         # TODO consider doing more correct checking of availability of properties/methods using native `` ticks
         property_name = property_name_for(method_name)